+++ mobile/android/android-components/components/browser/icons/src/main/java/mozilla/components/browser/icons/DesiredSize.kt
@@ -4,8 +4,19 @@
 
 package mozilla.components.browser.icons
 
+import androidx.annotation.Px
+
+/**
+ * Represents the desired size of an icon loaded by [BrowserIcons].
+ *
+ * @property targetSize The size the image will be displayed at, in pixels.
+ * @property maxSize The maximum size of an image before it will be thrown out, in pixels.
+ * Extremely large images are suspicious and should be ignored.
+ * @property maxScaleFactor The factor that the image can be scaled up before being thrown out.
+ * A lower scale factor results in less pixelation.
+ */
 data class DesiredSize(
-    val targetSize: Int,
-    val maxSize: Int,
+    @Px val targetSize: Int,
+    @Px val maxSize: Int,
     val maxScaleFactor: Float
 )