+++ mobile/android/android-components/components/feature/pwa/src/main/java/mozilla/components/feature/pwa/WebAppShortcutManager.kt
@@ -191,7 +191,7 @@ private fun webAppIcons(
         DataUriIconLoader()
     ),
     decoders = listOf(
-        AndroidIconDecoder(ignoreSize = true),
+        AndroidIconDecoder(),
         ICOIconDecoder()
     ),
     processors = listOf(