+++ mobile/android/android-components/docs/changelog.md
@@ -40,6 +40,9 @@ permalink: /changelog/
 * **lib-push-amazon**
   * 🆕 Added a new component for Amazon Device Messaging push support.
 
+* **browser-icons**
+  * Changed the maximum size for decoded icons. Icons are now scaled to the target size to save memory.
+
 # 5.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v4.0.0...v5.0.0)